moves kafka classes into their own folder
